modify from JID of incoming group MMS chat messages
Rather than just swapping <xmpp://+111111111> for <Contact> in the message
body, we strip it out completely and change the from JID to whoever
is in that contact. We're not modifying any of the MAM stuff, so this
won't mess up other clients; it's entirely within Dino's locally stored
message db.