don't check if the JID is in the roster before modifying the from JID
Previously, if a name wasn't in the roster, the from JID wouldn't be
re-written. This rewrites the from JID regardless of roster status.

This may or may not be a good idea; previous behavior would show
a message as being from "+11234561234,+19876544321@cheogram.com",
with the body starting with "<xmpp:+19876544321@cheogram.com>". The
"<xmpp:+19876544321@cheogram.com>" was clickable, though, so you could
easily set it in your roster.

The new behavior strips "<xmpp:+19876544321@cheogram.com>" from the
body completely, and shows the message as being from
"+19876544321@cheogram.com". But the "+19876544321@cheogram.com" is
not clickable, so you have to manually enter it (or cut & paste it)
to add it to the roster. So we may be reverting this... Or we may be
sending a separate patch to dino-im to make from JIDs clickable.